# openapi.json
{"openapi":"3.1.0","info":{"title":"Crypto Data API","description":"API for cryptocurrency funding rate and BTC dominance data","version":"1.0.0"},"paths":{"/crypto/funding_rate_historical/":{"get":{"tags":["funding_rate_historical"],"summary":"Get Funding Rate Controller","description":"Lấy dữ liệu lịch sử funding rate\n\nTham số:\n- symbols: Lấy nhiều hơn 2 mã giao dịch cách nhau bởi dấu phẩy (ví dụ: \"BTCUSDT,ETHUSDT\")\n- days: Số ngày cần lấy\n\nVí dụ: /crypto/funding_rate_historical/?symbols=BTCUSDT,ETHUSDT&days=7","operationId":"get_funding_rate_controller_crypto_funding_rate_historical__get","parameters":[{"name":"symbols","in":"query","required":false,"schema":{"type":"string","default":"BTCUSDT","title":"Symbols"}},{"name":"days","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FundingRateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/crypto/btc-dominance/":{"get":{"tags":["btc-dominance"],"summary":"Get Btc Dominance Data","description":"BTC Dominance Data\n\nTham số:\n- days: Số ngày cần lấy (0: realtime)\n\nVí dụ: /crypto/btc-dominance/?days=7","operationId":"get_btc_dominance_data_crypto_btc_dominance__get","parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BTCDominanceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/crypto/etf-candlestick/":{"get":{"tags":["etf-candlestick"],"summary":"Get Etf Candlestick Data","description":"ETF Candlestick data\n\n- symbol: Mã ETF (Có sẵn: E1VFVN30, FUEABVND, FUEBFVND, FUEDCMID, FUEFCV50, FUEIP100, FUEKIV30, FUEKIVFS, FUEKIVND, FUEMAV30, FUEMAVND, FUESSV30, FUESSV50, FUESSVFL, FUETCC50, FUEVFVND, FUEVN100)\n- day: Số ngày cần lấy (0=realtime)\n\nVí dụ: /crypto/etf-candlestick/?symbol=FUEVN100&day=7","operationId":"get_etf_candlestick_data_crypto_etf_candlestick__get","parameters":[{"name":"symbol","in":"query","required":false,"schema":{"type":"string","default":"FUEVN100","title":"Symbol"}},{"name":"day","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Day"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ETFCandlestickResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/crypto/gold-data/":{"get":{"tags":["gold-data"],"summary":"Get Gold Data","description":"Gold data\n\nTham số:\n- day: Số ngày cần lấy (0=realtime)\n\nVí dụ: /crypto/gold-data/?day=7","operationId":"get_gold_data_crypto_gold_data__get","parameters":[{"name":"day","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Day"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoldDataResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/crypto/funding_rate_realtime/":{"get":{"tags":["funding_rate_realtime"],"summary":"Get Realtime Funding Rate Controller","description":"realtime funding rate data\n\nTham số:\n- symbols: Lấy nhiều hơn 2 mã giao dịch thì viết cách nhau bởi dấu phẩy (ví dụ: \"BTCUSDT,ETHUSDT\")\n\nVí dụ: /crypto/funding_rate_realtime/?symbols=BTCUSDT,ETHUSDT","operationId":"get_realtime_funding_rate_controller_crypto_funding_rate_realtime__get","parameters":[{"name":"symbols","in":"query","required":false,"schema":{"type":"string","default":"BTCUSDT","title":"Symbols"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RealtimeFundingRateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/monitoring/funding-rate":{"get":{"tags":["crypto-monitoring"],"summary":"Check Funding Rate","description":"Kiểm tra dữ liệu funding rate theo chu kỳ 8h, 4h, 1h\n- 8h: Kiểm tra theo mốc thời gian 00:00, 08:00, 16:00\n- 4h: Kiểm tra theo mốc thời gian 00:00, 04:00, 08:00, 12:00, 16:00, 20:00\n- 1h: Kiểm tra mỗi giờ","operationId":"check_funding_rate_api_v1_monitoring_funding_rate_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Check Funding Rate Api V1 Monitoring Funding Rate Get"}}}}}}},"/api/v1/monitoring/btc-dominance":{"get":{"tags":["crypto-monitoring"],"summary":"Check Btc Dominance","description":"Check BTC Dominance có được cập nhật gần đây không","operationId":"check_btc_dominance_api_v1_monitoring_btc_dominance_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Check Btc Dominance Api V1 Monitoring Btc Dominance Get"}}}}}}},"/api/v1/monitoring/etf-candlestick":{"get":{"tags":["crypto-monitoring"],"summary":"Check Etf Candlestick","description":"Check ETF candle stick có được cập nhật gần đây không\n- Kiểm tra tất cả symbols ETF có dữ liệu gần đây không","operationId":"check_etf_candlestick_api_v1_monitoring_etf_candlestick_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Check Etf Candlestick Api V1 Monitoring Etf Candlestick Get"}}}}}}},"/":{"get":{"summary":"Root","description":"Health check endpoint","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Detailed health check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BTCDominanceResponse":{"properties":{"data":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Data"}},"type":"object","required":["data"],"title":"BTCDominanceResponse"},"ETFCandlestickModel":{"properties":{"symbol":{"type":"string","title":"Symbol"},"open":{"type":"number","title":"Open"},"high":{"type":"number","title":"High"},"low":{"type":"number","title":"Low"},"close":{"type":"number","title":"Close"},"volume":{"type":"number","title":"Volume"},"datetime":{"type":"string","title":"Datetime"}},"type":"object","required":["symbol","open","high","low","close","volume","datetime"],"title":"ETFCandlestickModel","description":"Model for ETF candlestick data"},"ETFCandlestickResponse":{"properties":{"data":{"items":{"$ref":"#/components/schemas/ETFCandlestickModel"},"type":"array","title":"Data"}},"type":"object","required":["data"],"title":"ETFCandlestickResponse","description":"Response model for historical ETF candlestick data"},"FundingRateResponse":{"properties":{"data":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Data"}},"type":"object","required":["data"],"title":"FundingRateResponse"},"GoldDataModel":{"properties":{"datetime":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Datetime"},"open":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Open"},"high":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"High"},"low":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Low"},"close":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Close"},"volume":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Volume"}},"type":"object","title":"GoldDataModel","description":"Model for gold minute data"},"GoldDataResponse":{"properties":{"data":{"items":{"$ref":"#/components/schemas/GoldDataModel"},"type":"array","title":"Data"}},"type":"object","required":["data"],"title":"GoldDataResponse","description":"Response model for gold data"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RealtimeFundingRate":{"properties":{"symbol":{"type":"string","title":"Symbol"},"funding_cap":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Funding Cap"},"funding_floor":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Funding Floor"},"funding_hour":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Funding Hour"},"funding_rate":{"type":"number","title":"Funding Rate"},"index_price":{"type":"number","title":"Index Price"},"interest_rate":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Interest Rate"},"interval":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Interval"},"mark_price":{"type":"number","title":"Mark Price"},"update_date":{"type":"string","title":"Update Date"},"update_time":{"type":"string","title":"Update Time"}},"type":"object","required":["symbol","funding_rate","index_price","mark_price","update_date","update_time"],"title":"RealtimeFundingRate"},"RealtimeFundingRateResponse":{"properties":{"data":{"items":{"$ref":"#/components/schemas/RealtimeFundingRate"},"type":"array","title":"Data"}},"type":"object","required":["data"],"title":"RealtimeFundingRateResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}